requirements.txt
Flask==3.0.0
gunicorn
python-dotenv==1.0.1
